+++ packages/graphql-hooks/src/middlewares/examples/debugMiddleware.js
@@ -1,0 +1,16 @@
+/**
+ * DebugMiddleware - example
+ * @param {function} logger
+ */
+const DebugMiddleware =
+  (logger = console.log) =>
+  (opts, next) => {
+    logger('Start request:', opts.operation)
+    opts.addResponseHook(res => {
+      logger('End request:', res)
+      return res
+    })
+    next()
+  }
+
+export default DebugMiddleware