+++ packages/graphql-hooks/src/useClientRequest.js
@@ -133,7 +133,6 @@ function useClientRequest(query, initialOpts = {}) {
   // arguments to fetchData override the useClientRequest arguments
   const fetchData = useDeepCompareCallback(
     newOpts => {
-      if (!isMounted.current) return Promise.resolve()
       const revisedOpts = {
         ...initialOpts,
         ...newOpts
@@ -143,6 +142,18 @@ function useClientRequest(query, initialOpts = {}) {
         ...operation,
         variables: revisedOpts.variables,
         operationName: revisedOpts.operationName
+      }
+
+      if (!isMounted.current) {
+        return Promise.resolve({
+          error: {
+            fetchError: new Error(
+              'fetchData should not be called after hook unmounted'
+            )
+          },
+          loading: false,
+          cacheHit: false
+        })
       }
 
       const revisedCacheKey = client.getCacheKey(revisedOperation, revisedOpts)
@@ -214,9 +225,10 @@ function useClientRequest(query, initialOpts = {}) {
   )
 
   // We perform caching after reducer update
-  // To include the outcome of updateData
+  // to include the outcome of updateData.
+  // The cache is already saved if in ssrMode.
   React.useEffect(() => {
-    if (state.useCache) {
+    if (state.useCache && !client.ssrMode) {
       client.saveCache(state.cacheKey, state)
     }
   }, [client, state])