+++ examples/fastify-ssr/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.5.1](https://github.com/nearform/graphql-hooks/compare/fastify-ssr@3.5.0...fastify-ssr@3.5.1) (2022-03-11)
+
+**Note:** Version bump only for package fastify-ssr
+
+
+
+
+
+# [3.5.0](https://github.com/nearform/graphql-hooks/compare/fastify-ssr@3.4.0...fastify-ssr@3.5.0) (2022-02-10)
+
+**Note:** Version bump only for package fastify-ssr
+
+
+
+
+
+# [3.4.0](https://github.com/nearform/graphql-hooks/compare/fastify-ssr@3.3.3...fastify-ssr@3.4.0) (2022-02-02)
+
+**Note:** Version bump only for package fastify-ssr
+
+
+
+
+
 ## [3.3.3](https://github.com/nearform/graphql-hooks/compare/fastify-ssr@3.3.2...fastify-ssr@3.3.3) (2022-01-04)
 
 **Note:** Version bump only for package fastify-ssr